+++ vision/detectors/__init__.py
@@ -1,0 +1,5 @@
+from .color import ColorDetector
+
+__all__ = [
+    "ColorDetector",
+]